Split daemon initialization lifecycle
