+++ config/max-lines-baseline.txt
@@ -11,7 +11,6 @@ inline src/main/claude-accounts/runtime-auth-service.ts
 inline src/main/codex-accounts/runtime-home-service.ts
 inline src/main/codex-accounts/service.ts
 inline src/main/codex/hook-service.ts
-inline src/main/daemon/daemon-init.ts
 inline src/main/git/worktree.ts
 inline src/main/github/project-view.ts
 inline src/main/index.ts
