+++ src/main/daemon/daemon-protocol-cleanup.ts
@@ -1,0 +1,98 @@
+import { unlinkSync } from 'node:fs'
+import { DaemonClient } from './client'
+import { DaemonEndpointOwnershipError } from './daemon-endpoint-adoption'
+import { probeDaemonSocket as probeSocket } from './daemon-launch-paths'
+import { getDaemonPidPath, getDaemonSocketPath, getDaemonTokenPath } from './daemon-spawner'
+import { killStaleDaemon } from './daemon-stale-kill'
+import { CLEAN_DISCONNECT_PROTOCOL_VERSION, type ListSessionsResult } from './types'
+
+const DAEMON_SELF_SHUTDOWN_WAIT_MS = 5_000
+
+export type OrphanedDaemonCleanupResult = {
+  /** True when a live daemon socket was found and torn down; false when none was running. */
+  cleaned: boolean
+  /** Number of live PTY sessions killed during cleanup (surfaced to the user). */
+  killedCount: number
+}
+
+export async function cleanupDaemonForProtocol(
+  runtimeDir: string,
+  protocolVersion: number
+): Promise<OrphanedDaemonCleanupResult> {
+  const socketPath = getDaemonSocketPath(runtimeDir, protocolVersion)
+  const tokenPath = getDaemonTokenPath(runtimeDir, protocolVersion)
+  const pidPath = getDaemonPidPath(runtimeDir, protocolVersion)
+
+  const alive = await probeSocket(socketPath)
+  if (!alive) {
+    if (protocolVersion >= CLEAN_DISCONNECT_PROTOCOL_VERSION) {
+      // Endpoint absence doesn't prove the PID record belongs to the current protocol; leave artifact cleanup to the owning daemon.
+      return { cleaned: false, killedCount: 0 }
+    }
+    try {
+      unlinkSync(pidPath)
+    } catch {
+      // Best-effort
+    }
+    return { cleaned: false, killedCount: 0 }
+  }
+
+  const client = new DaemonClient({ socketPath, tokenPath, protocolVersion })
+  let killedCount = 0
+  let didRequestShutdown = false
+  let didKillStaleDaemon = false
+  try {
+    await client.ensureConnected()
+    const sessions = await client
+      .request<ListSessionsResult>('listSessions', undefined)
+      .catch(() => ({ sessions: [] }))
+    killedCount = sessions.sessions.filter((s) => s.isAlive).length
+
+    // Use the single-shot `shutdown` RPC (kills all sessions then exits) to avoid racing per-session `kill` calls against the daemon exiting.
+    await client.request('shutdown', { killSessions: true }).catch(() => {
+      // Daemon exits immediately after the RPC, so the socket may close before the reply arrives; treat as success.
+    })
+    didRequestShutdown = true
+  } catch {
+    // Previous-protocol daemons may be wedged or too old for the RPC path; fall back to PID cleanup (only unlinks a live socket after proving the process is killed).
+    const killOutcome = await killStaleDaemon(runtimeDir, socketPath, tokenPath, protocolVersion)
+    didKillStaleDaemon = killOutcome.killed
+    if (killOutcome.liveOwnerSurvived) {
+      // Why: something still owns the endpoint. Returning as if it were cleaned lets restart
+      // fork a replacement that cannot publish onto the held name, leaving the user with no
+      // daemon instead of the one still running.
+      throw new DaemonEndpointOwnershipError(
+        'Daemon cleanup aborted: the existing daemon could not be confirmed stopped'
+      )
+    }
+  } finally {
+    client.disconnect()
+  }
+
+  if (didRequestShutdown && protocolVersion >= CLEAN_DISCONNECT_PROTOCOL_VERSION) {
+    if (!(await waitForDaemonEndpointExit(socketPath))) {
+      // Never fork a replacement while the old incarnation may still own the endpoint or be disposing terminal children.
+      throw new Error('Timed out waiting for daemon self-shutdown')
+    }
+    return { cleaned: true, killedCount }
+  }
+
+  try {
+    unlinkSync(pidPath)
+  } catch {
+    // Best-effort
+  }
+
+  return { cleaned: didRequestShutdown || didKillStaleDaemon, killedCount }
+}
+
+async function waitForDaemonEndpointExit(socketPath: string): Promise<boolean> {
+  const deadline = Date.now() + DAEMON_SELF_SHUTDOWN_WAIT_MS
+  while (Date.now() < deadline) {
+    if (!(await probeSocket(socketPath))) {
+      return true
+    }
+    await new Promise((resolve) => setTimeout(resolve, 50))
+  }
+  return !(await probeSocket(socketPath))
+}
