+++ src/main/daemon/daemon-provider-routing.ts
@@ -1,0 +1,29 @@
+import { DegradedDaemonPtyProvider } from './degraded-daemon-pty-provider'
+import type { DaemonPtyAdapter } from './daemon-pty-adapter'
+import { DaemonPtyRouter } from './daemon-pty-router'
+
+export type DaemonProvider = DaemonPtyRouter | DaemonPtyAdapter | DegradedDaemonPtyProvider
+
+export function getCurrentDaemonAdapter(provider: DaemonProvider): DaemonPtyAdapter {
+  if (provider instanceof DaemonPtyRouter || provider instanceof DegradedDaemonPtyProvider) {
+    return provider.getCurrentAdapter()
+  }
+  return provider
+}
+
+export function getLegacyDaemonAdapters(provider: DaemonProvider): DaemonPtyAdapter[] {
+  if (provider instanceof DaemonPtyRouter || provider instanceof DegradedDaemonPtyProvider) {
+    return [...provider.getLegacyAdapters()]
+  }
+  return []
+}
+
+export function disposeProviderSubscriptionsOnly(provider: DaemonProvider): void {
+  if (provider instanceof DaemonPtyRouter) {
+    provider.disposeRouterOnly()
+    return
+  }
+  if (provider instanceof DegradedDaemonPtyProvider) {
+    provider.disposeProviderOnly()
+  }
+}
